printing the size of c data types

+++ 0x00-hello_world/6-size.c
@@ -12,11 +12,11 @@ int main(void)
 	char char1;
 	long int num3;
 	float num4;
-	printf("Size of char: %lu byte(s)\n",sizeof(char1));
-	printf("Size of int: %lu byte(s)\n",sizeof(num1));
-	printf("Size of long int: %lu byte(s)\n",sizeof(num3));
-	printf("Size of long long int: %lu byte(s)\n",sizeof(num2));
-	printf("Size of float: %lu byte(s)\n",sizeof(num4));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(char1));
+	printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(num1));
+	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(num3));
+	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(num2));
+	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(num4));
 	return (0);
 
 }
